로그인 관련 권한 체크

+++ src/main/java/com/stj/tunnel/BlackLizard/controller/usr/ArticleController.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.stj.tunnel.BlackLizard.dto.Article;
 import com.stj.tunnel.BlackLizard.service.ArticleService;
@@ -30,14 +29,13 @@ public class ArticleController {
 		int totalPage = (int)Math.ceil(totalCount / (double)itemsCountInAPage);
 		int pageMenuArmSize = 5;
 		int page = Util.getAsInt(param.get("page"), 1);
-		int pageMenuStart = page - pageMenuArmSize;
 		
+		int pageMenuStart = page - pageMenuArmSize;
 		if (pageMenuStart < 1) {
 			pageMenuStart = 1;
 		}
 		
 		int pageMenuEnd = page + pageMenuArmSize;
-		
 		if (pageMenuEnd > totalPage) {
 			pageMenuEnd = totalPage;
 		}
@@ -84,14 +82,43 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("/usr/article/doWrite")
-	public String doWrite(@RequestParam Map<String, Object> param) {
+	public String doWrite(HttpSession session, Model model, @RequestParam Map<String, Object> param) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("redirectUri", "/usr/member/login");
+			
+			return "/common/redirect";
+		}
+		
 		int id = articleService.writeArticle(param);
 		
-		return String.format("<script> alert('%d번 게시물이 생성되었습니다.'); location.replace('/usr/article/detail?id=%d'); </script>", id, id);
+		model.addAttribute("msg", String.format("%d번 게시물이 생성되었습니다.", id));
+		model.addAttribute("redirectUri", String.format("/usr/article/detail?id=%d", id));
+		
+		return "/common/redirect";
 	}
 	
 	@RequestMapping("/usr/article/modify")
-	public String showModify(Model model, int id) {
+	public String showModify(HttpSession session, Model model, int id) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("redirectUri", "/usr/member/login");
+			
+			return "/common/redirect";
+		}
+		
 		Article article = articleService.getArticleById(id);
 		
 		model.addAttribute("article", article);
@@ -100,7 +127,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("/usr/article/doModify")
-	public String doModify(Model model, int id, String title, String body) {
+	public String doModify(HttpSession session, Model model, int id, String title, String body) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("redirectUri", "/usr/member/login");
+			
+			return "/common/redirect";
+		}
+				
 		articleService.modifyArticle(id, title, body);
 		
 		model.addAttribute("msg", String.format("%d번 게시물이 수정되었습니다.", id));
@@ -110,7 +150,20 @@ public class ArticleController {
 	}
 	
 	@RequestMapping("/usr/article/doDelete")
-	public String doDelete(Model model, int id) {
+	public String doDelete(HttpSession session, Model model, int id) {
+		int loginedMemberId = 0;
+		
+		if (session.getAttribute("loginedMemberId") != null) {
+			loginedMemberId = (int)session.getAttribute("loginedMemberId");
+		}
+		
+		if (loginedMemberId == 0) {
+			model.addAttribute("msg", "로그인 후 이용해주세요.");
+			model.addAttribute("redirectUri", "/usr/member/login");
+			
+			return "/common/redirect";
+		}
+		
 		articleService.deleteArticleById(id);
 		
 		model.addAttribute("msg", String.format("%d번 게시물이 삭제되었습니다.", id));
